+++ example/included_modifiers.dart
@@ -1,0 +1,27 @@
+/// Flutter library that transform common-use widgets into companion modifiers,
+/// countering tree-like structure in code.
+/// Please refer to https://github.com/CW-Henry/widget_modifiers for
+/// documentation about the library.
+enum IncludedModifiers {
+  /// Interactions Modifiers
+  gesture,
+  ripple,
+
+  /// Layout Modifiers
+  frame,
+  align,
+  center,
+  aspectRatio,
+  expand,
+
+  /// Painting Modifiers
+  border,
+  shadow,
+  transform,
+  opacity,
+
+  /// Styling Modifiers
+  padding,
+}
+
+