+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0
+
+* Stable release 
+* Categorized and improved modifiers, added example session.
 ## 0.1.0
 
-Implemented basic modifier toolsets for common widgets
+* Implemented basic modifier toolsets for common widgets